Refactor: - This change is part of a larger refactoring effort to streamline the server code and improve maintainability.
