+++ src/metadata_parser.cpp
@@ -2,10 +2,13 @@
 #include <iostream>
 #include <regex>
 #include <cstdio>
-#include "server_bbox.hpp" // Include the new server-side BBox definition
+#include <chrono>
+#include <queue>
+#include <arpa/inet.h>
+#include <openssl/ssl.h>
+#include "ssl.hpp"
 
-// tcp_server.hpp에서 선언된 update_bbox_buffer 함수 사용
-void update_bbox_buffer(const std::vector<ServerBBox>& new_bboxes);
+
 
 using namespace std;
 
@@ -13,6 +16,12 @@ using namespace std;
 vector<ServerBBox> latest_bboxes;
 mutex bbox_mutex;
 atomic<bool> parser_running(false);
+
+// BBox 버퍼링 관련 전역 변수
+std::atomic<int> bbox_buffer_delay_ms(2400);   // 기본값 2.4초
+std::atomic<int> bbox_send_interval_ms(50);   // 기본값 50ms
+std::queue<TimestampedBBox> bbox_buffer;
+std::mutex bbox_buffer_mutex;
 
 void start_metadata_parser() {
     parser_running = true;
@@ -119,4 +128,134 @@ void parse_metadata() {
     }
 
     pclose(pipe);
+}
+
+// ==================== BBox 버퍼 관리 함수들 ====================
+
+void clear_bbox_buffer() {
+    std::lock_guard<std::mutex> lock(bbox_buffer_mutex);
+    
+    // 버퍼의 모든 데이터 제거
+    std::queue<TimestampedBBox> empty_queue;
+    bbox_buffer.swap(empty_queue);
+    
+    std::cout << "[BUFFER] Buffer cleared completely" << std::endl;
+}
+
+void update_bbox_buffer(const std::vector<ServerBBox>& new_bboxes) {
+    std::lock_guard<std::mutex> lock(bbox_buffer_mutex);
+    
+    // 새로운 bbox 데이터를 타임스탬프와 함께 버퍼에 저장
+    TimestampedBBox data{
+        std::chrono::steady_clock::now(),
+        new_bboxes
+    };
+    bbox_buffer.push(data);
+    
+    // 버퍼 크기 관리 (너무 많은 데이터가 쌓이지 않도록)
+    auto now = std::chrono::steady_clock::now();
+    int removed_count = 0;
+    
+    // 1. 10초보다 오래된 데이터 제거
+    while (!bbox_buffer.empty()) {
+        const auto& oldest = bbox_buffer.front();
+        auto age = std::chrono::duration_cast<std::chrono::milliseconds>(
+            now - oldest.timestamp).count();
+            
+        if (age > 10000) {  // 10초보다 오래된 데이터 제거
+            bbox_buffer.pop();
+            removed_count++;
+        } else {
+            break;
+        }
+    }
+    
+    // 2. 버퍼 크기가 너무 크면 강제로 오래된 데이터 제거 (최대 50개 유지)
+    while (bbox_buffer.size() > 50) {
+        bbox_buffer.pop();
+        removed_count++;
+    }
+    
+    // 디버그 로그
+    if (removed_count > 0 || bbox_buffer.size() > 10) {
+        std::cout << "[BUFFER] Added 1 item, removed " << removed_count 
+                 << " items, current size: " << bbox_buffer.size() << std::endl;
+    }
+}
+
+bool send_bboxes_to_client(SSL* ssl) {
+    std::vector<ServerBBox> bboxes_to_send;
+    int buffer_size = 0;
+    int processed_count = 0;
+    
+    {
+        std::lock_guard<std::mutex> lock(bbox_buffer_mutex);
+        
+        auto now = std::chrono::steady_clock::now();
+        buffer_size = bbox_buffer.size();
+        
+        // 버퍼에서 지연 시간이 지난 가장 오래된 데이터 찾기
+        while (!bbox_buffer.empty()) {
+            const auto& oldest = bbox_buffer.front();
+            auto age = std::chrono::duration_cast<std::chrono::milliseconds>(
+                now - oldest.timestamp).count();
+                
+            if (age >= bbox_buffer_delay_ms.load()) {
+                bboxes_to_send = oldest.bboxes;
+                bbox_buffer.pop();
+                processed_count++;
+                break;  // 하나만 처리하고 나감
+            }
+            break;  // 지연 시간이 지나지 않은 데이터는 아직 전송하지 않음
+        }
+        
+        // 디버그 로그 추가
+        if (buffer_size > 5) {  // 버퍼에 5개 이상 쌓이면 경고
+            std::cout << "[DEBUG] Buffer getting large: " << buffer_size 
+                     << " items, processed: " << processed_count << std::endl;
+        }
+    }
+    
+    if (bboxes_to_send.empty()) {
+        return true;  // 전송할 데이터가 없음 (에러 아님)
+    }
+
+    nlohmann::json bbox_array = nlohmann::json::array();
+    for (const ServerBBox& box : bboxes_to_send) {
+        nlohmann::json j = {
+            {"id", box.object_id},
+            {"type", box.type},
+            {"confidence", box.confidence},
+            {"x", box.left},
+            {"y", box.top},
+            {"width", box.right - box.left},
+            {"height", box.bottom - box.top}
+        };
+        bbox_array.push_back(j);
+    }
+    
+    nlohmann::json response = {
+        {"response_id", 200},
+        {"bboxes", bbox_array},
+        {"buffer_info", {
+            {"buffer_size", buffer_size},
+            {"processed_count", processed_count}
+        }}
+    };
+
+    std::string json_str = response.dump();
+    uint32_t res_len = json_str.length();
+    uint32_t net_res_len = htonl(res_len);
+
+    // 먼저 4바이트 길이 접두사 전송
+    if (sendAll(ssl, reinterpret_cast<const char*>(&net_res_len), sizeof(net_res_len), 0) == -1) {
+        std::cout << "[TCP Server] Failed to send length prefix" << std::endl;
+        return false;
+    }
+    // 그 다음 실제 JSON 데이터 전송
+    int ret = sendAll(ssl, json_str.c_str(), res_len, 0);
+    if (ret == -1) {
+        std::cout << "[TCP Server] Failed to send JSON data" << std::endl;
+    }
+    return ret != -1;
 }
