+++ src/tcp_server.hpp
@@ -1,4 +1,4 @@
-// TCP 서버 구동 모듈
+// TCP 서버 구동 모듈 (리팩토링된 버전)
 
 #pragma once
 
@@ -50,41 +50,41 @@
 // 비밀번호 / 복구코드 해싱을 위한 헤더 파일
 #include "hash.hpp"
 
-// BBox 버퍼링을 위한 구조체
-struct TimestampedBBox {
-    std::chrono::steady_clock::time_point timestamp;
-    std::vector<ServerBBox> bboxes;
-};
-
-// 전역 변수 선언
-extern std::atomic<int> bbox_buffer_delay_ms;  // 버퍼 지연 시간 (M ms)
-extern std::atomic<int> bbox_send_interval_ms; // 전송 주기 (N ms)
-extern std::queue<TimestampedBBox> bbox_buffer;
-extern std::mutex bbox_buffer_mutex;
-
 using namespace std;
 using json = nlohmann::json;
 
 const int PORT = 8080;
 
+// ==================== 기존 함수들 ====================
 string getLines();
-
 string putLines(CrossLine);
-
 string deleteLines(int index);
-
-string base64_encode(const vector<unsigned char>& in);
-
 int tcp_run();
 
-bool recvAll(SSL*, char* buffer, size_t len);
+// ==================== 새로운 유틸리티 함수들 ====================
+SSL* setup_ssl_connection(int client_socket);
+void initialize_database_tables(SQLite::Database& db);
+bool receive_json_message(SSL* ssl, json& received_json);
 
-ssize_t sendAll(SSL*, const char* buffer, size_t len, int flags);
+// ==================== 요청 처리 함수들 ====================
+void handle_detection_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_line_insert_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_line_select_all_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_line_delete_all_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_baseline_insert_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_vertical_line_insert_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_baseline_select_all_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_login_step1_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_login_step2_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_signup_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex);
+void handle_bbox_start_request(SSL* ssl, std::atomic<bool>& bbox_push_enabled, std::thread& push_thread, std::thread& metadata_thread);
+void handle_bbox_stop_request(SSL* ssl, std::atomic<bool>& bbox_push_enabled, std::thread& push_thread, std::thread& metadata_thread);
 
-void printNowTimeKST();
+// ==================== 요청 라우터 및 정리 함수들 ====================
+void route_request(SSL* ssl, const json& received_json, SQLite::Database& db, std::mutex& db_mutex, 
+                   std::atomic<bool>& bbox_push_enabled, std::thread& push_thread, std::thread& metadata_thread);
+void cleanup_client_connection(SSL* ssl, int client_socket, std::atomic<bool>& bbox_push_enabled, 
+                               std::thread& push_thread, std::thread& metadata_thread);
 
-bool send_bboxes_to_client(SSL* ssl);
-
-// BBox 버퍼 관리 함수
-void update_bbox_buffer(const std::vector<ServerBBox>& new_bboxes);
-void clear_bbox_buffer();
+// ==================== 메인 클라이언트 처리 함수 ====================
+void handle_client(int client_socket, SQLite::Database& db, std::mutex& db_mutex);
